+++ libsolidity/ast/ASTAnnotations.h
@@ -237,12 +237,8 @@ struct IdentifierPathAnnotation: TypeNameAnnotation
 {
 	/// Referenced declaration, set during reference resolution stage.
 	Declaration const* referencedDeclaration = nullptr;
-};
-
-struct UserDefinedTypeNameAnnotation: TypeNameAnnotation
-{
-	/// Referenced declaration, set during reference resolution stage.
-	Declaration const* referencedDeclaration = nullptr;
+	/// What kind of lookup needs to be done (static, virtual, super) find the declaration.
+	SetOnce<VirtualLookup> requiredLookup;
 };
 
 struct ExpressionAnnotation: ASTAnnotation